+++ anxcloud/schema_kubernetes.go
@@ -81,6 +81,16 @@ Enable autoscaling for this cluster. Defaults to false if unset.
 			Optional: true,
 			ForceNew: true,
 		},
+		"apiserver_allowlist": {
+			Type:        schema.TypeList,
+			Description: "A list of CIDRs that should be allowed access to the kubernetes API server. By default there are no IP restrictions.",
+			Optional:    true,
+			Computed:    true,
+			ForceNew:    true,
+			Elem: &schema.Schema{
+				Type: schema.TypeString,
+			},
+		},
 	}
 }